Refactor analytics pipeline to support one data type per pipeline
