+++ crates/sui-analytics-indexer/src/analytics_metrics.rs
@@ -2,62 +2,38 @@
 // SPDX-License-Identifier: Apache-2.0
 #![allow(dead_code)]
 
-use axum::{extract::Extension, http::StatusCode, routing::get, Router};
-use mysten_metrics::RegistryService;
-use prometheus::{
-    register_int_counter_with_registry, register_int_gauge_with_registry, IntCounter, IntGauge,
-    Registry, TextEncoder,
-};
 use std::net::SocketAddr;
+
+use axum::{extract::Extension, http::StatusCode, routing::get, Router};
+use prometheus::{
+    register_int_counter_vec_with_registry, register_int_gauge_vec_with_registry, IntCounterVec,
+    IntGaugeVec, Registry, TextEncoder,
+};
+
+use mysten_metrics::RegistryService;
 
 const METRICS_ROUTE: &str = "/metrics";
 
 #[derive(Clone)]
 pub struct AnalyticsMetrics {
-    pub total_checkpoint_received: IntCounter,
-    pub total_transaction_received: IntCounter,
-    pub total_transaction_object_received: IntCounter,
-    pub total_object_received: IntCounter,
-    pub total_event_received: IntCounter,
-    pub last_uploaded_checkpoint: IntGauge,
+    pub total_received: IntCounterVec,
+    pub last_uploaded_checkpoint: IntGaugeVec,
 }
 
 impl AnalyticsMetrics {
     pub fn new(registry: &Registry) -> Self {
         Self {
-            total_checkpoint_received: register_int_counter_with_registry!(
-                "total_checkpoint_received",
-                "Total number of checkpoints received",
-                registry,
+            total_received: register_int_counter_vec_with_registry!(
+                "total_received",
+                "Number of checkpoints received",
+                &["data_type"],
+                registry
             )
             .unwrap(),
-            total_transaction_received: register_int_counter_with_registry!(
-                "total_transaction_received",
-                "Total number of transactions received",
-                registry,
-            )
-            .unwrap(),
-            total_transaction_object_received: register_int_counter_with_registry!(
-                "total_transaction_object_received",
-                "Total number of transaction objects received",
-                registry,
-            )
-            .unwrap(),
-            total_object_received: register_int_counter_with_registry!(
-                "total_object_received",
-                "Total number of objects received",
-                registry,
-            )
-            .unwrap(),
-            total_event_received: register_int_counter_with_registry!(
-                "total_event_received",
-                "Total number of events received",
-                registry,
-            )
-            .unwrap(),
-            last_uploaded_checkpoint: register_int_gauge_with_registry!(
+            last_uploaded_checkpoint: register_int_gauge_vec_with_registry!(
                 "last_uploaded_checkpoint",
-                "Last checkpoint uploaded to remote store",
+                "Number of uploaded checkpoints.",
+                &["data_type"],
                 registry,
             )
             .unwrap(),
