+++ crates/sui-core/src/db_checkpoint_handler.rs
@@ -213,7 +213,7 @@ impl DBCheckpointHandler {
     async fn upload_db_checkpoints_to_object_store(&self, missing_epochs: Vec<u64>) -> Result<()> {
         let last_missing_epoch = missing_epochs.last().cloned().unwrap_or(0);
         let local_checkpoints_by_epoch =
-            find_all_dirs_with_epoch_prefix(&self.input_object_store).await?;
+            find_all_dirs_with_epoch_prefix(&self.input_object_store, None).await?;
         let mut dirs: Vec<_> = local_checkpoints_by_epoch.iter().collect();
         dirs.sort_by_key(|(epoch_num, _path)| *epoch_num);
         for (epoch, db_path) in dirs {
@@ -255,7 +255,7 @@ impl DBCheckpointHandler {
     }
     async fn garbage_collect_old_db_checkpoints(&self) -> Result<Vec<u64>> {
         let local_checkpoints_by_epoch =
-            find_all_dirs_with_epoch_prefix(&self.input_object_store).await?;
+            find_all_dirs_with_epoch_prefix(&self.input_object_store, None).await?;
         let mut deleted = Vec::new();
         for (epoch, path) in local_checkpoints_by_epoch.iter() {
             let marker_paths: Vec<Path> = self
@@ -336,7 +336,8 @@ mod tests {
             false,
         )?;
         let local_checkpoints_by_epoch =
-            find_all_dirs_with_epoch_prefix(&db_checkpoint_handler.input_object_store).await?;
+            find_all_dirs_with_epoch_prefix(&db_checkpoint_handler.input_object_store, None)
+                .await?;
         assert!(!local_checkpoints_by_epoch.is_empty());
         assert_eq!(*local_checkpoints_by_epoch.first_key_value().unwrap().0, 0);
         assert_eq!(
